upd views & .MD

+++ Makefile
@@ -1,12 +1,9 @@
 all: init-db install run
-# 	sqlite3 shop.db < init.sql
-# 	npm install
-# 	npm start
+
 init-db:
-# 	if [ -f shop.db ]; then
-# 		rm shop.db
-# 	fi
-	rm shop.db
+	if [ -f shop.db ]; then
+		rm shop.db
+	fi
 	sqlite3 shop.db < init.sql
 install:
 	npm install
